results/rebuttal/run-20231111-034248_LLMatic-zero_shot_donnemartin-system-design-oop-finance_0_20-runs_log/finance/17/LLMatic/user.py
import hashlib

# Mock database
users_db = {}


def hash_password(password):
	return hashlib.sha256(password.encode()).hexdigest()


def create_user(username, password, email):
	if username in users_db:
		return 'Username already exists'
	else:
		users_db[username] = {'password': hash_password(password), 'email': email, 'bank_accounts': []}
		return 'User created successfully'


def login(username, password):
	if username in users_db and users_db[username]['password'] == hash_password(password):
		return 'Login successful'
	else:
		return 'Invalid username or password'


def recover_password(username):
	if username in users_db:
		return 'Password recovery email sent'
	else:
		return 'Username does not exist'


def link_bank_account(username, bank_name, account_number, balance):
	if username in users_db:
		users_db[username]['bank_accounts'].append({'bank_name': bank_name, 'account_number': account_number, 'balance': balance, 'transactions': []})
		return 'Bank account linked successfully'
	else:
		return 'Username does not exist'


def import_transactions(username, account_number, transactions):
	if username in users_db:
		for account in users_db[username]['bank_accounts']:
			if account['account_number'] == account_number:
				account['transactions'].extend(transactions)
				account['balance'] = sum([transaction['amount'] for transaction in transactions])
				return 'Transactions imported successfully'
	return 'Username or account number does not exist'


def update_account_balance(username, account_number, amount):
	if username in users_db:
		for account in users_db[username]['bank_accounts']:
			if account['account_number'] == account_number:
				account['balance'] += amount
				return 'Account balance updated successfully'
	return 'Username or account number does not exist'


def get_savings_tips(username):
	if username in users_db:
		# Analyze the user's transactions and provide personalized tips for saving money
		# This is a mock implementation for the purpose of this task
		return ['Tip 1', 'Tip 2', 'Tip 3']
	else:
		return 'Username does not exist'


def get_product_recommendations(username):
	if username in users_db:
		# Recommend financial products based on the user's profile
		# This is a mock implementation for the purpose of this task
		return ['Product 1', 'Product 2', 'Product 3']
	else:
		return 'Username does not exist'

